JSONInventorySlipsWeb-main/requirements.txt
flask==3.1.1
flask-session==0.5.0
pandas==2.3.0
docxtpl==0.20.0
python-docx==1.1.2
docxcompose==1.4.0
werkzeug==3.1.3
gunicorn==21.2.0